requirements.txt
aiogram[fast]
httpx
sqlalchemy
pydantic
python-dotenv
asyncpg
aiohttp